requirements.txt
discord.py==2.4.0
Pillow==11.1.0
pymongo==4.10.1
python-dotenv==1.0.1
